move WithOrSkip to utils
